+++ ruler-frontend/src/main/kotlin/com/spotify/ruler/frontend/Main.kt
@@ -16,35 +16,33 @@
 
 package com.spotify.ruler.frontend
 
-import com.spotify.ruler.frontend.components.report
+import com.spotify.ruler.frontend.components.Report
 import com.spotify.ruler.models.AppReport
-import kotlinext.js.require
 import kotlinx.browser.document
-import kotlinx.html.dom.append
-import kotlinx.html.js.link
 import kotlinx.serialization.decodeFromString
 import kotlinx.serialization.json.Json
-import react.Props
-import react.createElement
-import react.dom.client.createRoot
+import org.jetbrains.compose.web.renderComposable
+
 
 fun main() {
-    require("./style.css")
-
-    require("bootstrap/dist/css/bootstrap.css")
-    require("bootstrap/dist/js/bootstrap.bundle.js")
+    kotlinext.js.require("./style.css")
+    kotlinext.js.require("bootstrap/dist/css/bootstrap.css")
+    kotlinext.js.require("bootstrap/dist/js/bootstrap.bundle.js")
 
     // Load and show the favicon
-    val favicon = require("./favicon.svg").toString()
-    document.head?.append?.link(href = favicon, rel = "icon")
+    val favicon = kotlinext.js.require("./favicon.svg").toString()
+    val link = document.createElement("link").apply {
+        setAttribute("rel", "icon")
+        setAttribute("href", favicon)
+    }
+    document.head?.append(link)
 
     // Load and deserialize the report data
-    val rawReport = require("report.json").toString()
+    val rawReport = kotlinext.js.require("report.json").toString()
     val report = Json.decodeFromString<AppReport>(rawReport)
 
     // Visualize and display the report data
-    val root = createRoot(requireNotNull(document.getElementById("root")))
-    root.render(createElement<Props> {
-        report(report)
-    })
+    renderComposable(rootElementId = "root") {
+        Report(report)
+    }
 }